Basics of menus handled
