+++ Let Ava Game/Assets/Prefabs/GameState.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &1812841194360869390
+--- !u!1 &7030103040027608160
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,8 +8,8 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 1812841194360869388}
-  - component: {fileID: 1812841194360869389}
+  - component: {fileID: 4537168899940762387}
+  - component: {fileID: 5219958875717257178}
   m_Layer: 0
   m_Name: GameState
   m_TagString: Untagged
@@ -17,30 +17,30 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &1812841194360869388
+--- !u!4 &4537168899940762387
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1812841194360869390}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_GameObject: {fileID: 7030103040027608160}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
   m_LocalPosition: {x: 978.1329, y: 420.66083, z: -33988.637}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1812841194360869389
+--- !u!114 &5219958875717257178
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1812841194360869390}
+  m_GameObject: {fileID: 7030103040027608160}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e4104a0a91b9a82449d34498632ce9fe, type: 3}
+  m_Script: {fileID: 11500000, guid: ba35ce0346ffc284f87c04a1a78ed9c6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   coinScore: 0
