--- Let Ava Game/Assets/Scenes/Nature_assets.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d38f3b6ddd7390f44b717e1122bb426a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
