--- Let Ava Game/Assets/Scenes/Nature_props_showcase.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6bbe50710129d0149ab951136c5115d0
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
