--- Let Ava Game/Assets/Scripts/Button_Handler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b1f8dd790c180274287ec4a92557c419
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
