+++ Let Ava Game/Assets/Scripts/Coin.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class Coin : MonoBehaviour
 {
     public float speed;
-    private SFXPlayer _SFXPlayer;
-    private GameState _gameState;
+    //private SFXPlayer _SFXPlayer;
+    //private GameState _gameState;
 
     void Start() {
-        _SFXPlayer = GameObject.Find("SFXPlayer").GetComponent<SFXPlayer>();
-        _gameState = GameObject.Find("GameState").GetComponent<GameState>();
+        //_SFXPlayer = GameObject.Find("SFXPlayer").GetComponent<SFXPlayer>();
+        //_gameState = GameObject.Find("GameState").GetComponent<GameState>();
     }
 
     void Update() {
@@ -25,14 +25,14 @@ public class Coin : MonoBehaviour
     void OnTriggerEnter2D(Collider2D other) {
         //Disappears when hits the player
         if (other.CompareTag("Player")) {
-            _SFXPlayer.playCoinCollect();
-            _gameState.coinScore += 1;
+            //_SFXPlayer.playCoinCollect();
+            //_gameState.coinScore += 1;
             Destroy(gameObject);
         }
 
         //Is is destroyed by projectiles
         if (other.CompareTag("Projectile")) {
-            _SFXPlayer.playCollectDestroy();
+            //_SFXPlayer.playCollectDestroy();
             Destroy(gameObject);
         }
     }
