--- Let Ava Game/Assets/Scripts/GameOver_Menu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6f7f5465af7a6fb4fb5d91eb376aa80d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
