+++ Let Ava Game/Assets/Scripts/GameState.cs
@@ -7,10 +7,4 @@ public class GameState : MonoBehaviour
     public float coinScore = 0;
     public float distance = 0;
     public float time = 0;
-    
-    void Update()
-    {;
-        distance += 1.5f * Time.deltaTime;
-        time += 1.0f * Time.deltaTime;
-    }
 }
