--- Let Ava Game/Assets/Scripts/Pause_Menu.cs
@@ -1,32 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Pause_Menu : MonoBehaviour
-{
-    public static bool isPaused = false;
-    public GameObject pauseMenuUI;
-
-    // Update is called once per frame
-    void Update() {
-        if (Input.GetKeyDown(KeyCode.Escape)) {
-            if (isPaused) {
-                Resume();
-            } else {
-                Pause();
-            }
-        }
-    }
-
-    public void Resume() {
-        pauseMenuUI.SetActive(false);
-        Time.timeScale = 1f;
-        isPaused = false;
-    }
-
-    void Pause() {
-        pauseMenuUI.SetActive(true);
-        Time.timeScale = 0f;
-        isPaused = true;
-    }
-}
