+++ Let Ava Game/Assets/Scripts/Scene/MainGame.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 96664a9d257b55d46854f0ab2a6e1e3f
+guid: 7951b61d25379c34f8e33e05a47046f9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
